Updates the style to match the reqquirements of our customer.
This required both the login and register page to be recreated so that
each element appeared on individual lines. I.e. with the titles above
the input boxes, Customer also requested links to move to login from
register and register from login, this has also been added in this
commit.

This commit closes ticket #31