Changed various styling of pages as instructed by the customer
The Login title on the login page has been moved to the left and
capitalised.
There is whitespace between the LOGIN and REGISTER buttons and the
sentences below them.
The Register title has now been captitalised.
There is more whitespace at the bottom of the Register page.

These fixes are related to issue #69 in the main repo.