can now see the items added in the wishlist page
we have added the ability to get the user from the wishlist page using
the previous cookie code that does the same thing and now we can access
the product and user data that we need to display on the wishlist. As a
placeholder, we have used the productlist component to display the
products on the wishlist at the moment for proof of concept.

This commit is in relation to issue #82 on the main repo