Added kids shoe filter, size and database elements as well as product page elements
To the database I added the field kidsSize this is used to hold kids sizes when main size is set to 0. Also changed the product page to show the kids shoe size with the prefix 'Kids' where the size field is 0. Also added the All options to size and condition.