Created footer and updated login and register
Created the footer that appears at the bottom of each page excluding the menu page, this has three parts, information which contains the about us, ts&cs and gdpr page and a connect with us that contains the socials as seen in menu with the addittion of linkedin as requested by cmr, these two sections are created as a table to keep the design uniform. Under this we have a copyright disclaimer. I also updated the login and register pages to remove the extra spacing that was added here that doesnt appear in the rest of the site. Also added the linkedIn SVG.